created curatednews.html
This contains our new Curated News headlines feature. The news selection and methodology has been vetted and decentralized (as much as possible) into a mostly manipulation-proof news solution. Obviously, from a security standpoint and a substantive one, our applications are a better option. However, there needed to be a way for desktop users to gain access to our features in non-obtrusive ways. It serves as part of a free preview with future updates forthcoming, including a rating structure.